Install Polly.js dependencies
These packages are designed to stub and record HTTP requests, whose payloads and response can be reused then to mock `fetch` requests in tests.

Since our project is wrapping an HTTP API, we are going to need a tool like this to write tests which are as close as possible to the production environment.