Setup Polly.js configuration with Jest
Now FetchClient uses Polly.js to record real HTTP responses and validate its corresponding assertions against the recorded HTTP responses.

This way we have a way of testing the client with actual data, and without the risk of breaking the test if at any point the test API changes/goes down.

On top of that we have also configured Polly.js to work with our current Jest environment, and defined a set of helpers to make easier to record HTTP requests in the future. With the `setupHttpRecording` method we set before and after individual test run callbacks to activate and activate HTTP request recording, keeping the tests clean.