Add CI lint job
This job runs the `package.json` `lint` script, which uses `ESLint` underneath.

To speed up the process I added a `--cache` option. This option generates a file with linter results, so in subsequent jobs only new/modified files are checked.

Our cache key is configured with the package-json and eslintrc configuration files, so whenever they are updated the cache is regenerated.