fix: HttpClient Body handling
We initially assumed a Body could only be a `Record` or a `string`. This is wrong, there are other possible formats such as `Buffer` or `Array` buffer.

On top of that, certain kinds of `body` require some serialization to be delivered in requests.

This commit defines a new `Body` class which handles different kinds of body payloads, pre-processing them accordingly to the client needs and providing a simple interface for providing the payloads serialized. Now our Http clients make use of this object to send requests with body.