chore: remove Fetch client
After struggling so much with the Fetch client I decided to remove it.

`fetch` seems to present a lot of problems with testing tools in Node environment. Reading around, I am not the only one facing problems, and it seems there are not so many fixes being released.

On top of that, the reMarkable API does not support communication through browser due to limitations with CORS security. This means we cannot use this library in the browser, but a node environment. Hence, from now on we will only rely on a single HTTP client, the Node https one.