chore: Add logic for merging to `HttpContext`.
It is very common when using HttpClient instances to specify new contexts (headers) to perform specific requests.

To better handle this logic I turned the HttpClientContext type into a class. It now presents a `merge` method which allows combining two `Contexts` into a single one.

When two `Contexts` are merge, host and header values are combined. If both context share a host or specific headers, the original values will be overwritten by the ones from the given context.