chore: refactor FileManager into FileBuffer
Rather than following the approach of having Service-like objects I decided to instead define more understandable concept-like classes which model the whole upload logic.

The `Buffer` is the payload we send to the API to upload files. It contains information about its own kind and know by itself how to upload.

Compared to the previous Manager approach, this feels more natural and makes it easier to handle different kinds of file types.