feature: HashUrl logic
the reMarkable API uses hashes (special strings) to identify any piece of information stored in the cloud (files, folders, metadata, notebooks, etc).

To access the hash content, the reMarkable API provides a `download` endpoint. This endpoint requests a hash and returns in exchange a signed URL which points to where the hash data is stored.

The `HashUrl` class encapsulates the logic to fetch the signed URL, given a hash, and download its content.