Add script to regenerate test data
The script, executable via `yarn test:setup`, provides a guided way to regenerate HTTP Polly records with a new token and delete this token once all tests are passing.

Since the recorder will include the token in the payloads, we need to make sure these tokens are invalidated after running the tests. The script provides and step-by-step guide to do so.

There are still some improvements to be made, since the file upload tests break some of the environment variables (so they should be executed in last place), but for now the solution is good enough to keep our HTTP records safe of valid token when submitting them to the repository.