Updated Typescript & EsLint configuration to ignore null-undefined related.
Disable strict type checking for Null / Undefined types and some ESLint rules related to coalescing values.

Definitely this is something we should better handle in our code, but for now it is just causing a lot of issues when checking types. I will disable it for the moment and come back to it in future PRs.