fix: CI pipeline. The pipeline uses ESLint with a cache strategy for speeding up style checks.
Due to the changes in the `lint` script this was broken. Now the pipeline calls both linters separatedly.

This is not the optimal solution. If in the future we want to add new style checks, we will have to manually write them in the CI. In a future PR we will address a solution to run linters with cache.