chore: separate FileBuffer tests from the rest when running tests.
When files are uploaded to reMarkable, folder hashes change. This breaks some of the credentials our tests are currently using to verify we are fetching the right data from the system.

In a different PR we should do some maintenance to better handle these situations but for now I separated the FileBuffer tests from the rest, so when re-generating HTTP records we avoid these side effects.