refactor: ServiceManager production endpoint method.
One of my goals refactoring is to remove unnecessary dependencies from the `authentication` namespace.

`HttpClients` should be dispatched from the `ServiceManager`, so other classes do not know how to initialize a client.

The `Device` class was instantiating its own instances of the class to perform authentication requests.

In this commit we add a new service endpoint for the `ServiceManager` which provides a `HttpClient` instance the `Device` can use to authenticate agains the reMarkable Cloud API.