refactor: `Device` `pair` token handling.
The pair `jwt` token holds a `Device` information. Thus, to instantiate a `Device` we just need its token. Here I replace the `Device` initialization logic to just make use of the pair token to fill the `Device` data.

I added a new class, `RemarkableTokenPayload`, capable of parsing a reMarkable JWT token and retrieve the relevant information. This class will replace later on the `DeviceToken` one which is unnecessary (redundant considering the `Device` class already holds the device information) and incorrect (since we are also using `DeviceToken` to handle session tokens).