refactor: `Device` `Session`.
Move the reMarkable cloud session handling logic to its own class and remove the now obsolete `DeviceToken`.

Now we have a `Session` class which properly encapsulates the session handling logic. It is still pretty simple, but it is going to help us handle more complex logic in the future (such as disconnecting a device from the cloud).

With this change the authentication module is finally refactored.