feat: File System Snapshot.
Added a new `internal` class to take snapshots of the reMarkable FileSystem.

The idea behind is for the `FileSystem` to always perform HTTP requests to the API to fetch the most up to date version of the FileSystem elements; and the snapshot to provide the FileSystem cache.

This way we do not need to worry about adding any sort of complex cache-like logic in our code. It is up to the developers how do they want to make use of the snapshots as a cache for optimizing operations.